+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * Option to exclude some layers from being rendered / exported
     * Freehand drawing tool
     * Interactive "add node" with the edit tool
+    * Bezier molding, click and drag on a curve to edit it, without using handles
 * Bug Fixes:
     * Exporting the document no longer alters the document save state
     * Fixed crash on Lottie export